pyproject.toml
[project]
name = "ai-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.2.0",
    "openai>=2.22.0",
    "python-dotenv>=1.2.1",
]
